Small tool for merging multiple glossary files.
Use case: the authors of a book want to combine this glossary with some terms they define themselves. Using this tool, they can merge `glosario/glossary.yml` and `book/extra-terms.yml`.

The tool is written in Python, but is meant to be used from the command line, not loaded as a module. (If I could have done it as a shell script, I would have...)