Update PULL_REQUEST_TEMPLATE.md
Despite @TomKellyGenetics's recent changes to the template, it is still not being used as we intend and we're all still getting spammed with notifications.

This way, the handles are all hidden in an html comment and the correct ones must be actively assigned. I've seen this done in lots of other repos that use PR templates.